Refactor heuristic agents into dedicated modules
