+++ gomoku/ai/heuristic/__init__.py
@@ -1,0 +1,13 @@
+"""ヒューリスティックエージェントを集めたサブパッケージ"""
+
+from .random_agent import RandomAgent
+from .immediate_win_block_agent import ImmediateWinBlockAgent
+from .four_three_priority_agent import FourThreePriorityAgent
+from .longest_chain_agent import LongestChainAgent
+
+__all__ = [
+    "RandomAgent",
+    "ImmediateWinBlockAgent",
+    "FourThreePriorityAgent",
+    "LongestChainAgent",
+]
